Squashapps/api/requirements.txt
alembic==0.9.8
aniso8601==3.0.0
attrs==19.1.0
bcrypt==3.1.4
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.4
Flask-Migrate==2.1.1
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
honcho==1.0.1
idna==2.8
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
psycopg2-binary==2.8.2
pycparser==2.18
pyrsistent==0.15.2
python-crontab==2.3.6
python-dateutil==2.7.0
python-editor==1.0.3
pytz==2018.3
pypng==0.0.20
PyQRCode==1.2.1
redis==3.2.1
requests==2.21.0
semantic-version==2.6.0
six==1.11.0
smmap2==2.0.5
SQLAlchemy==1.2.5
urllib3==1.24.1
virtualenv==16.4.3
Werkzeug==0.14.1
xlrd==1.2.0
elasticsearch==7.0.2
zeep